pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "black==24.10.0",
    "fastapi>=0.135.2",
    "flake8==7.0.0",
    "httpx==0.26.0",
    "isort>=8.0.1",
    "prometheus-client>=0.25.0",
    "pymongo>=4.16.0",
    "pytest==8.0.0",
    "pytest-asyncio==0.21.2",
    "pytest-cov==4.1.0",
    "python-dotenv>=1.2.2",
    "python-json-logger>=4.1.0",
    "requests>=2.33.0",
    "slowapi>=0.1.9",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
